Move all non-start .py files to non-root directories
- Finished root directory cleanup by moving everything in it to either main_window or com_window.
- Switched imports to use relative paths where possible.
