+++ start.py
@@ -1,10 +1,7 @@
-import main
+from com_window.main import *
 
-window = None
-
-
-def start(*args, **kwargs):  # TODO: does this need its own file??
-    global window
-
-    window = main.MainApp(*args, **kwargs)
+if __name__ == "__main__":
+    app = QtWidgets.QApplication(sys.argv)
+    window = comWindow()
     window.show()
+    sys.exit(app.exec_())
